Graphics.UI.WX

Description

The WX module just re-exports functionality from helper modules and defines the start function.

The WX library provides a haskellized interface to the raw wxWidgets functionality provided by the Graphics.UI.WXCore library.

Functions

start :: IO a -> IO ()

start runs the GUI.

Modules

module Graphics.UI.WX.Types

module Graphics.UI.WX.Attributes

module Graphics.UI.WX.Classes

module Graphics.UI.WX.Variable

module Graphics.UI.WX.Layout

module Graphics.UI.WX.Events

module Graphics.UI.WX.Window

module Graphics.UI.WX.Frame

module Graphics.UI.WX.Timer

module Graphics.UI.WX.Media

module Graphics.UI.WX.Menu

module Graphics.UI.WX.Controls

module Graphics.UI.WX.Dialogs

module Graphics.UI.WX.Draw