Graphics.UI.WX.Controls

Description

Defines common GUI controls.

Classes

data Align

Alignment.

Constructors

AlignLeft

AlignRight

AlignCentre

class Aligned w where

Widgets that can have aligned content. Note: this property is not used to set the alignment of a widget itself -- See Graphics.UI.WXCore.Layout for more information about layout.

Minimal complete definition

alignment

Methods

alignment :: CreateAttr w Align

Set the alignment of the content. Due to wxWidgets constrictions, this property has to be set at creation time.

alignment :: Aligned w => CreateAttr w Align

Set the alignment of the content. Due to wxWidgets constrictions, this property has to be set at creation time.

data Wrap

Wrap mode.

Constructors

WrapNone
No wrapping (and show a horizontal scrollbar).

WrapLine
Wrap lines that are too long at any position.

WrapWord
Wrap lines that are too long at word boundaries.

class Wrapped w where

Widgets that have wrappable content.

Minimal complete definition

wrap

Methods

wrap :: CreateAttr w Wrap

Set the wrap mode of a widget.

wrap :: Wrapped w => CreateAttr w Wrap

Set the wrap mode of a widget.

class Sorted w where

Widgets that have sorted contents.

Minimal complete definition

sorted

Methods

sorted :: CreateAttr w Bool

Is the content of the widget sorted?

sorted :: Sorted w => CreateAttr w Bool

Is the content of the widget sorted?

Calendar Ctrl

calendarCtrl :: Window a -> [Prop (CalendarCtrl ())] -> IO (CalendarCtrl ())

date :: (Typeable a, IsDate a) => Attr (CalendarCtrl w) a

class IsDate a where

Minimal complete definition

toWXDate, fromWXDate

Methods

toWXDate :: a -> IO (DateTime ())

fromWXDate :: DateTime () -> IO a

Containers

type Panel a = Window (CPanel a)

Pointer to an object of type Panel, derived from Window.

panel :: Window a -> [Prop (Panel ())] -> IO (Panel ())

Create a Panel, a window that is normally used as a container for controls. It has a standard background and maintains standard keyboard navigation (ie. Tab moves through the controls).

Note: the defaultButton attribute is removed. Set defaultButton to parent Frame or Dialog instead of this control now. This is an incompatible change to support wxWidgets 2.8.x.

Attributes: focusOn

Instances: Form -- Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Textual, Literate, Reactive, Paint

panelEx :: Window a -> Style -> [Prop (Panel ())] -> IO (Panel ())

Create a Panel with a specific style.

Note: the defaultButton attribute is removed. Set defaultButton to parent Frame or Dialog instead of this control now. This is an incompatible change to support wxWidgets 2.8.x.

Attributes: focusOn

Instances: Form -- Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Textual, Literate, Reactive, Paint

type Notebook a = Control (CNotebook a)

Pointer to an object of type Notebook, derived from Control.

notebook :: Window a -> [Prop (Notebook ())] -> IO (Notebook ())

Create a Notebook. Layout is managed with the tabs combinator.

Instances: Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Textual, Literate, Reactive, Paint

focusOn :: Window a -> IO ()

Set the initial focus on this control.

Controls

Button

type Button a = Control (CButton a)

Pointer to an object of type Button, derived from Control.

button :: Window a -> [Prop (Button ())] -> IO (Button ())

Create a standard push button.

Instances: Commanding -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

buttonEx :: Window a -> Style -> [Prop (Button ())] -> IO (Button ())

Create a standard push button with the given flags.

Instances: Commanding -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

smallButton :: Window a -> [Prop (Button ())] -> IO (Button ())

Create a minimially sized push button.

Instances: Commanding -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

buttonRes :: Window a -> String -> [Prop (Button ())] -> IO (Button ())

Complete the construction of a push button instance which has been loaded from a resource file.

Instances: Commanding -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

type BitmapButton a = Button (CBitmapButton a)

Pointer to an object of type BitmapButton, derived from Button.

bitmapButton :: Window a -> [Prop (BitmapButton ())] -> IO (BitmapButton ())

Create a bitmap button. Use the picture attribute to set the bitmap.

Instances: Commanding, Pictured -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

bitmapButtonRes :: Window a -> String -> [Prop (BitmapButton ())] -> IO (BitmapButton ())

Complete the construction of a bitmap button instance which has been loaded from a resource file.

Instances: Commanding, Pictured -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

Text entry

type TextCtrl a = Control (CTextCtrl a)

Pointer to an object of type TextCtrl, derived from Control.

entry :: Window a -> [Prop (TextCtrl ())] -> IO (TextCtrl ())

Create a single-line text entry control. Note: alignment has to be set at creation time (or the entry has default alignment (=left) ). This is an alias for textEntry

Instances: Wrap, Aligned, Commanding, Updating -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

textEntry :: Window a -> [Prop (TextCtrl ())] -> IO (TextCtrl ())

Create a single-line text entry control. Note: alignment can only be set at creation time (default is left alignment).

Instances: Wrap, Aligned, Commanding, Updating -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

textCtrl :: Window a -> [Prop (TextCtrl ())] -> IO (TextCtrl ())

Create a multi-line text control. Note: the wrap and alignment can only be set at creation time, the defaults are WrapNone and AlignLeft respectively.

Instances: Wrap, Aligned, Commanding, Updating -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

textCtrlRich :: Window a -> [Prop (TextCtrl ())] -> IO (TextCtrl ())

Create a multi-line text rich-text control with a certain wrap mode Enables font and color settings on windows, while being equal to textCtrl on other platforms. Note: the wrap and alignment can only be set at creation time, the defaults are WrapNone and AlignLeft respectively.

Instances: Wrap, Aligned, Commanding, Updating -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

textCtrlEx :: Window a -> Style -> [Prop (TextCtrl ())] -> IO (TextCtrl ())

Create a generic text control given a certain style.

Instances: Wrap, Aligned, Commanding, Updating -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

textCtrlRes :: Window a -> String -> [Prop (TextCtrl ())] -> IO (TextCtrl ())

Complete the construction of a text control instance which has been loaded from a resource file.

Instances: Wrap, Aligned, Commanding, Updating -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

processEnter :: Styled w => Attr w Bool

Process enter key events, used in a comboBox or textCtrl and catched using an on command handler (otherwise pressing Enter is either processed internally by the control or used for navigation between dialog controls).

processTab :: Styled w => Attr w Bool

Process tab key events, used in a comboBox or textCtrl. (otherwise pressing Tab is either processed internally by the control or used for navigation between dialog controls).

CheckBox

type CheckBox a = Control (CCheckBox a)

Pointer to an object of type CheckBox, derived from Control.

checkBox :: Window a -> [Prop (CheckBox ())] -> IO (CheckBox ())

Create a new checkbox.

Instances: Commanding,Checkable -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

checkBoxRes :: Window a -> String -> [Prop (CheckBox ())] -> IO (CheckBox ())

Complete the construction of a check box instance which has been loaded from a resource file.

Instances: Commanding,Checkable -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

Choice

type Choice a = Control (CChoice a)

Pointer to an object of type Choice, derived from Control.

choice :: Window a -> [Prop (Choice ())] -> IO (Choice ())

Create a choice item to select one of a list of strings.

Instances: Sorted, Selecting,Selection,Items -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

choiceEx :: Window a -> Style -> [Prop (Choice ())] -> IO (Choice ())

Create a choice item, given a set of style flags, to select one of a list of strings

Instances: Selecting,Selection,Items -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

choiceRes :: Window a -> String -> [Prop (Choice ())] -> IO (Choice ())

Complete the construction of a choice instance which has been loaded from a resource file.

ComboBox

type ComboBox a = Choice (CComboBox a)

Pointer to an object of type ComboBox, derived from Choice.

comboBox :: Window a -> [Prop (ComboBox ())] -> IO (ComboBox ())

Create a new combo box.

Instances: Selecting, Commanding, Updating, Selection, Items -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

A command event is triggered when the enter key is pressed and when processEnter has been set to True.

comboBoxEx :: Window a -> Style -> [Prop (ComboBox ())] -> IO (ComboBox ())

Create a new combo box with a given set of flags.

Instances: Selecting, Commanding, Updating, Selection, Items -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

A command event is triggered when the enter key is pressed and when processEnter has been set to True.

comboBoxRes :: Window a -> String -> [Prop (ComboBox ())] -> IO (ComboBox ())

Complete the construction of a combo box instance which has been loaded from a resource file.

ListBox

type ListBox a = Control (CListBox a)

Pointer to an object of type ListBox, derived from Control.

type SingleListBox a = ListBox (CSingleListBox a)

Pointer to single selection list boxes, deriving from ListBox.

type MultiListBox a = ListBox (CMultiListBox a)

Pointer to multiple selection list boxes, deriving from ListBox.

singleListBox :: Window a -> [Prop (SingleListBox ())] -> IO (SingleListBox ())

Create a single selection list box.

Instances: Sorted,Selecting,Selection,Items -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

singleListBoxRes :: Window a -> String -> [Prop (SingleListBox ())] -> IO (SingleListBox ())

Complete the construction of a single list box instance which has been loaded from a resource file.

multiListBox :: Window a -> [Prop (MultiListBox ())] -> IO (MultiListBox ())

multiListBoxRes :: Window a -> String -> [Prop (MultiListBox ())] -> IO (MultiListBox ())

Complete the construction of a single list box instance which has been loaded from a resource file.

data ListBoxView b a

A small wrapper over WX's ListCtrl, allowing us to keep the data we're representing as well as its string form (shown to the user as rows).

Constructors

ListBoxView

Fields

listBoxViewCtrl :: ListBox b

listBoxViewItems :: Var [a]

listBoxViewToRow :: a -> String

singleListBoxView :: Window b -> [Prop (SingleListBox ())] -> (a -> String) -> IO (ListBoxView (CSingleListBox ()) a)

multiListBoxView :: Window b -> [Prop (MultiListBox ())] -> (a -> String) -> IO (ListBoxView (CMultiListBox ()) a)

listBoxViewAddItem :: ListBoxView b a -> a -> IO ()

listBoxViewGetItems :: ListBoxView b a -> IO [a]

listBoxViewSetItems :: ListBoxView b a -> [a] -> IO ()

singleListBoxViewGetSelection :: ListBoxView (CSingleListBox ()) a -> IO (Maybe a)

multiListBoxViewGetSelections :: ListBoxView (CMultiListBox ()) a -> IO [a]

RadioBox

type RadioBox a = Control (CRadioBox a)

Pointer to an object of type RadioBox, derived from Control.

radioBox :: Window a -> Orientation -> [String] -> [Prop (RadioBox ())] -> IO (RadioBox ())

Create a new radio button group with an initial orientation and a list of labels. Use selection to get the currently selected item.

Instances: Selecting,Selection,Items -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

radioBoxRes :: Window a -> String -> [Prop (RadioBox ())] -> IO (RadioBox ())

Complete the construction of a radio box instance which has been loaded from a resource file.

Spin Control

type SpinCtrl a = Control (CSpinCtrl a)

Pointer to an object of type SpinCtrl, derived from Control.

spinCtrl :: Window a -> Int -> Int -> [Prop (SpinCtrl ())] -> IO (SpinCtrl ())

Create a spin control: a text field with up/down buttons. The value (selection) is always between a specified minimum and maximum.

Instances: Selection, Selecting -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

spinCtrlRes :: Window a -> String -> [Prop (SpinCtrl ())] -> IO (SpinCtrl ())

Complete the construction of a spin control instance which has been loaded from a resource file.

Slider

type Slider a = Control (CSlider a)

Pointer to an object of type Slider, derived from Control.

hslider :: Window a -> Bool -> Int -> Int -> [Prop (Slider ())] -> IO (Slider ())

Create a horizontal slider with a specified minimum and maximum. Set the Bool argument to True to show labels (minimum, maximum, and current value). The selection attribute gives the current value.

Instances: Commanding,Selection -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

vslider :: Window a -> Bool -> Int -> Int -> [Prop (Slider ())] -> IO (Slider ())

Create a vertical slider with a specified minimum and maximum. Set the Bool argument to True to show labels (minimum, maximum, and current value). The selection attribute gives the current value.

Instances: Commanding,Selection -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

sliderEx :: Window a -> Int -> Int -> Style -> [Prop (Slider ())] -> IO (Slider ())

Create a slider with a specified minimum and maximum. The selection attribute gives the current value.

Instances: Commanding,Selection -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

sliderRes :: Window a -> String -> [Prop (Slider ())] -> IO (Slider ())

Complete the construction of a slider instance which has been loaded from a resource file.

Gauge

type Gauge a = Control (CGauge a)

Pointer to an object of type Gauge, derived from Control.

hgauge :: Window a -> Int -> [Prop (Gauge ())] -> IO (Gauge ())

Create a horizontal gauge with a specified integer range (max value). The selection attribute determines the position of the gauge.

Instances: Selection -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

vgauge :: Window a -> Int -> [Prop (Gauge ())] -> IO (Gauge ())

Create a vertical gauge with a specified integer range (max value). The selection attribute determines the position of the gauge.

Instances: Selection -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

gaugeEx :: Window a -> Int -> Style -> [Prop (Gauge ())] -> IO (Gauge ())

Create a gauge control. The selection attribute determines the position of the gauge.

Instances: Selection -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

gaugeRes :: Window a -> String -> [Prop (Gauge ())] -> IO (Gauge ())

Complete the construction of a gauge instance which has been loaded from a resource file.

ToggleButton

type ToggleButton a = Control (CToggleButton a)

Pointer to an object of type ToggleButton, derived from Control.

type BitmapToggleButton a = ToggleButton (CBitmapToggleButton a)

Pointer to an object of type BitmapToggleButton, derived from ToggleButton.

toggleButton :: Window a -> [Prop (ToggleButton ())] -> IO (ToggleButton ())

Create a toggle button.

Instances: Commanding, Checkable, -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

bitmapToggleButton :: Window a -> [Prop (BitmapToggleButton ())] -> IO (BitmapToggleButton ())

Create a bitmap toggle button. Use the picture attribute to set the bitmap.

Instances: Commanding, Pictured -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

Tree control

type TreeCtrl a = Control (CTreeCtrl a)

Pointer to an object of type TreeCtrl, derived from Control.

treeCtrl :: Window a -> [Prop (TreeCtrl ())] -> IO (TreeCtrl ())

Create a single-selection tree control with buttons (i.e. + and - signs).

Attributes: treeEvent

Instances: Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

treeCtrlEx :: Window a -> Style -> [Prop (TreeCtrl ())] -> IO (TreeCtrl ())

Create a tree control.

Attributes: treeEvent

Instances: Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

treeEvent :: Event (TreeCtrl a) (EventTree -> IO ())

Tree control events.

treeCtrlRes :: Window a -> String -> [Prop (TreeCtrl ())] -> IO (TreeCtrl ())

Complete the construction of a tree control instance which has been loaded from a resource file.

List control

type ListCtrl a = Control (CListCtrl a)

Pointer to an object of type ListCtrl, derived from Control.

listCtrl :: Window a -> [Prop (ListCtrl ())] -> IO (ListCtrl ())

Create a report-style list control.

Attributes: listEvent, columns

Instances: Items -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

listCtrlEx :: Window a -> Style -> [Prop (ListCtrl ())] -> IO (ListCtrl ())

Create a list control.

Attributes: listEvent, columns

Instances: Items -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

listCtrlRes :: Window a -> String -> [Prop (ListCtrl ())] -> IO (ListCtrl ())

Complete the construction of a list control instance which has been loaded from a resource file.

listCtrlSetColumnWidths :: ListCtrl () -> Int -> IO ()

listEvent :: Event (ListCtrl a) (EventList -> IO ())

List control events.

columns :: Attr (ListCtrl a) [(String, Align, Int)]

The columns attribute controls the columns in a report-view list control.

data ListView a

A small wrapper over WX's ListCtrl, allowing us to keep the data we're representing as well as its string form (shown to the user as rows).

Constructors

ListView

Fields

listViewCtrl :: ListCtrl ()

listViewItems :: Var [a]

listViewToRow :: a -> [String]

listViewLayout :: ListView a -> Layout

listViewSetHandler :: ListView a -> (EventList -> IO ()) -> IO ()

listViewSelectHandle :: ListView a -> (Maybe a -> IO ()) -> EventList -> IO ()

listViewSetItems :: ListView a -> [a] -> IO ()

listViewGetItems :: ListView a -> IO [a]

listViewAddItem :: ListView a -> a -> IO ()

listView :: Window b -> [String] -> (a -> [String]) -> IO (ListView a)

Static text

type StaticText a = Control (CStaticText a)

Pointer to an object of type StaticText, derived from Control.

staticText :: Window a -> [Prop (StaticText ())] -> IO (StaticText ())

Create static text label, see also label.

staticTextRes :: Window a -> String -> [Prop (StaticText ())] -> IO (StaticText ())

Complete the construction of a static text label instance which has been loaded from a resource file.

SplitterWindow

type SplitterWindow a = Window (CSplitterWindow a)

Pointer to an object of type SplitterWindow, derived from Window.

splitterWindow :: Window a -> [Prop (SplitterWindow ())] -> IO (SplitterWindow ())

Create a splitter window.

Instances: Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

ImageList

type ImageList a = WxObject (CImageList a)

Pointer to an object of type ImageList, derived from WxObject.

imageList :: Size -> IO (ImageList ())

Create an empty image list that will contain images of the desired size.

imageListFromFiles :: Size -> [FilePath] -> IO (ImageList ())

Create an image list containing the images in the supplied file name list that will be scaled towards the desired size.

MediaCtrl

data MediaCtrlBackend

Optional backend for your MediaCtrl. If you want to know more about backend, you must see wxWidgets' document: http://docs.wxwidgets.org/trunk/classwx_media_ctrl.html#mediactrl_choosing_backend

Constructors

DirectShow
Use ActiveMovie/DirectShow. Default backend on Windows.

MediaControlInterface
Use Media Command Interface. Windows Only.

WindowsMediaPlayer10
Use Windows Media Player 10. Windows Only. Require to use wxWidgets 2.8.x.

QuickTime
Use QuickTime. Mac Only.

GStreamer
Use GStreamer. Unix Only. Require GStreamer and GStreamer Support.

DefaultBackend
Use default backend on your platform.

type MediaCtrl a = Window (CMediaCtrl a)

Pointer to an object of type MediaCtrl, derived from Window.

mediaCtrl :: Window a -> [Prop (MediaCtrl ())] -> IO (MediaCtrl ())

mediaCtrlWithBackend :: Window a -> MediaCtrlBackend -> [Prop (MediaCtrl ())] -> IO (MediaCtrl ())

Create MediaCtrl with choosing backend. This is useful to select backend on Windows. But if you don't want to cause any effect to other platforms, you must use wxToolkit or #ifdef macro to choose correct function for platforms. For example,

import Graphics.UI.WXCore.Defines
...
  m <- case wxToolkit of
         WxMSW -> mediaCtrlWithBackend f MediaControlInterface []
         _     -> mediaCtrl f []

or

#ifdef mingw32_HOST_OS || mingw32_TARGET_OS
  m <-  mediaCtrlWithBackend f MediaControlInterface []
#else
  m <-  mediaCtrl f []
#endif

mediaCtrlEx :: Window a -> Style -> MediaCtrlBackend -> [Prop (MediaCtrl ())] -> IO (MediaCtrl ())

Wizard

type Wizard a = Dialog (CWizard a)

Pointer to an object of type Wizard, derived from Dialog.

wizard :: Window a -> [Prop (Wizard ())] -> IO (Wizard ())

Create an empty wizard.

wizardEx :: Window a -> Style -> [Prop (Wizard ())] -> IO (Wizard ())

wizardPageSimple :: Wizard a -> [Prop (WizardPageSimple ())] -> IO (WizardPageSimple ())

Create an empty simple wizard page.

runWizard :: Wizard a -> WizardPage b -> IO Bool

Run the wizard. IMPORTANT: windowDestroy must be called on the wizard when it is no longer used. After windowDestroy has been called, the wizard or any of its children must not be accessed anymore.

next :: WriteAttr (WizardPageSimple a) (Maybe (WizardPageSimple b))

prev :: WriteAttr (WizardPageSimple a) (Maybe (WizardPageSimple b))

chain :: [WizardPageSimple a] -> IO ()

Chain together all given wizard pages.

wizardPageSize :: Attr (Wizard a) Size

wizardEvent :: Event (Wizard a) (EventWizard -> IO ())

wizardCurrentPage :: ReadAttr (Wizard a) (Maybe (WizardPage ()))

StyledTextCtrl

type StyledTextCtrl a = Control (CStyledTextCtrl a)

Pointer to an object of type StyledTextCtrl, derived from Control.

stcEvent :: Event (StyledTextCtrl ()) (EventSTC -> IO ())

styledTextCtrl :: Window a -> [Prop (StyledTextCtrl ())] -> IO (StyledTextCtrl ())

styledTextCtrlEx :: Window a -> Style -> [Prop (StyledTextCtrl ())] -> IO (StyledTextCtrl ())

PropertyGrid

type PropertyGrid a = Control (CPropertyGrid a)

Pointer to an object of type PropertyGrid, derived from Control.

propertyGrid :: Window a -> [Prop (PropertyGrid ())] -> IO (PropertyGrid ())

Create a property grid.

Attributes: propertyGridEvent,

Instances: Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

propertyGridEvent :: Event (PropertyGrid a) (EventPropertyGrid -> IO ())

PropertyGrid control events.

Orphan instances

Form (Panel a)

Methods

layout :: Attr (Panel a) Layout

Selection (Slider a)

Methods

selection :: Attr (Slider a) Int

Selection (Gauge a)

Methods

selection :: Attr (Gauge a) Int

Selection (SpinCtrl a)

Methods

selection :: Attr (SpinCtrl a) Int

Selection (RadioBox a)

Methods

selection :: Attr (RadioBox a) Int

Selection (Choice ())

Methods

selection :: Attr (Choice ()) Int

Selection (ComboBox a)

Methods

selection :: Attr (ComboBox a) Int

Checkable (ToggleButton a)

Methods

checkable :: Attr (ToggleButton a) Bool

checked :: Attr (ToggleButton a) Bool

Checkable (CheckBox a)

Methods

checkable :: Attr (CheckBox a) Bool

checked :: Attr (CheckBox a) Bool

Pictured (BitmapToggleButton a)

Methods

picture :: Attr (BitmapToggleButton a) FilePath

Pictured (BitmapButton a)

Methods

picture :: Attr (BitmapButton a) FilePath

Updating (TextCtrl a)

Methods

update :: Event (TextCtrl a) (IO ())

Updating (ComboBox a)

Methods

update :: Event (ComboBox a) (IO ())

Commanding (Slider a)

Methods

command :: Event (Slider a) (IO ())

Commanding (ToggleButton a)

Methods

command :: Event (ToggleButton a) (IO ())

Commanding (TextCtrl a)

Methods

command :: Event (TextCtrl a) (IO ())

Commanding (Button a)

Methods

command :: Event (Button a) (IO ())

Commanding (ComboBox a)

Methods

command :: Event (ComboBox a) (IO ())

Commanding (CheckBox a)

Methods

command :: Event (CheckBox a) (IO ())

Selecting (SpinCtrl a)

Methods

select :: Event (SpinCtrl a) (IO ())

Selecting (RadioBox a)

Methods

select :: Event (RadioBox a) (IO ())

Selecting (ListBox a)

Methods

select :: Event (ListBox a) (IO ())

Selecting (Choice ())

Methods

select :: Event (Choice ()) (IO ())

Selecting (ComboBox a)

Methods

select :: Event (ComboBox a) (IO ())

Media (MediaCtrl a)

Methods

play :: MediaCtrl a -> IO ()

stop :: MediaCtrl a -> IO ()

Items (RadioBox a) String

Methods

itemCount :: ReadAttr (RadioBox a) Int

items :: Attr (RadioBox a) [String]

item :: Int -> Attr (RadioBox a) String

itemDelete :: RadioBox a -> Int -> IO ()

itemsDelete :: RadioBox a -> IO ()

itemAppend :: RadioBox a -> String -> IO ()

Items (ListBox a) String

Methods

itemCount :: ReadAttr (ListBox a) Int

items :: Attr (ListBox a) [String]

item :: Int -> Attr (ListBox a) String

itemDelete :: ListBox a -> Int -> IO ()

itemsDelete :: ListBox a -> IO ()

itemAppend :: ListBox a -> String -> IO ()

Items (Choice a) String

Methods

itemCount :: ReadAttr (Choice a) Int

items :: Attr (Choice a) [String]

item :: Int -> Attr (Choice a) String

itemDelete :: Choice a -> Int -> IO ()

itemsDelete :: Choice a -> IO ()

itemAppend :: Choice a -> String -> IO ()

Items (ListCtrl a) [String]

Methods

itemCount :: ReadAttr (ListCtrl a) Int

items :: Attr (ListCtrl a) [[String]]

item :: Int -> Attr (ListCtrl a) [String]

itemDelete :: ListCtrl a -> Int -> IO ()

itemsDelete :: ListCtrl a -> IO ()

itemAppend :: ListCtrl a -> [String] -> IO ()