Graphics.UI.WX.Events

Description

Define event handling. Events are parametrised by the widget that can correspond to a certain event and the type of the event handler. For example, the resize event has type:

Reactive w => Event w (IO ())

This means that all widgets in the Reactive class can respond to resize events (and since Window is an instance of this class, this means that basically all visible widgets are reactive).

An Event w a can be transformed into an attribute of type Attr w a using the on function.

do f <- frame [text := "test"]
   set f [on resize := set f [text := "resizing"]]

For convenience, the mouse and keyboard have a serie of event filters: click, drag, enterKey, charKey, etc. These filters are write-only and do not overwrite any previous mouse or keyboard handler but all stay active at the same time. However, all filters will be overwritten again when mouse or keyboard is set again. For example, the following program makes sense:

set w [on click := ..., on drag := ...]

But in the following program, only the handler for mouse will be called:

set w [on click := ..., on mouse := ...]

If you want to set the mouse later but retain the old event filters, you can first read the current mouse handler and call it in the new handler (and the same for the keyboard, of course). This implementation technique is used to implement event filters themselves and is also very useful when setting an event handler for a closing event:

set w [on closing :~ \previous -> do{ ...; previous }]

Note that you should call propagateEvent (or skipCurrentEvent) whenever you do not process the event yourself in an event handler. This propagates the event to the parent event handlers and give them a chance to handle the event in an appropiate way. This gives another elegant way to install a closing event handler:

set w [on closing := do{ ...; propagateEvent }]

Event

data Event w a

An event for a widget w that expects an event handler of type a.

on :: Event w a -> Attr w a

Transform an event to an attribute.

mapEvent :: (a -> b) -> (a -> b -> a) -> Event w a -> Event w b

Change the event type.

propagateEvent :: IO ()

Pass the event on the next wxWidgets event handler, either on this window or its parent. Always call this method when you do not process the event. (This function just call skipCurrentEvent).

Basic events

Selecting

class Selecting w where

Selecting widgets fire a select event when an item is selected.

Minimal complete definition

select

Methods

select :: Event w (IO ())

A select event is fired when an item is selected.

select :: Selecting w => Event w (IO ())

A select event is fired when an item is selected.

Commanding

class Commanding w where

Commanding widgets fire a command event.

Minimal complete definition

command

Methods

command :: Event w (IO ())

A commanding event, for example a button press.

command :: Commanding w => Event w (IO ())

A commanding event, for example a button press.

Updating

class Updating w where

Updating widgets fire an update event.

Minimal complete definition

update

Methods

update :: Event w (IO ())

An update event, for example when the text of a TextCtrl changes. An update event, unlike a command event, is typically a passive state change and doesn't require a response.

update :: Updating w => Event w (IO ())

An update event, for example when the text of a TextCtrl changes. An update event, unlike a command event, is typically a passive state change and doesn't require a response.

Reactive

class Reactive w where

Reactive widgets are almost all visible widgets on the screen.

Minimal complete definition

mouse, keyboard, closing, idle, resize, focus, activate

Methods

mouse :: Event w (EventMouse -> IO ())

keyboard :: Event w (EventKey -> IO ())

closing :: Event w (IO ())

idle :: Event w (IO Bool)

resize :: Event w (IO ())

focus :: Event w (Bool -> IO ())

activate :: Event w (Bool -> IO ())

mouse :: Reactive w => Event w (EventMouse -> IO ())

keyboard :: Reactive w => Event w (EventKey -> IO ())

closing :: Reactive w => Event w (IO ())

idle :: Reactive w => Event w (IO Bool)

resize :: Reactive w => Event w (IO ())

focus :: Reactive w => Event w (Bool -> IO ())

activate :: Reactive w => Event w (Bool -> IO ())

class Paint w where

Paint widgets can serve as a canvas. Note: it is illegal to use both a paint and paintRaw event handler at the same widget.

Minimal complete definition

paint, paintRaw, paintGc, repaint

Methods

paint :: Event w (DC () -> Rect -> IO ())

Paint double buffered to a device context. The context is always cleared before drawing. Takes the current view rectangle (adjusted for scrolling) as an argument.

paintRaw :: Event w (PaintDC () -> Rect -> [Rect] -> IO ())

Paint directly to the on-screen device context. Takes the current view rectangle and a list of dirty rectangles as arguments.\

paintGc :: Event w (GCDC () -> Rect -> IO ())

Paint double buffered to a GCDC device context, for anti-aliased drawing. The context is always cleared before drawing. Takes the current view rectangle (adjusted for scrolling) as an argument.

repaint :: w -> IO ()

Emit a paint event to the specified widget.

paint :: Paint w => Event w (DC () -> Rect -> IO ())

Paint double buffered to a device context. The context is always cleared before drawing. Takes the current view rectangle (adjusted for scrolling) as an argument.

paintRaw :: Paint w => Event w (PaintDC () -> Rect -> [Rect] -> IO ())

Paint directly to the on-screen device context. Takes the current view rectangle and a list of dirty rectangles as arguments.\

paintGc :: Paint w => Event w (GCDC () -> Rect -> IO ())

Paint double buffered to a GCDC device context, for anti-aliased drawing. The context is always cleared before drawing. Takes the current view rectangle (adjusted for scrolling) as an argument.

repaint :: Paint w => w -> IO ()

Emit a paint event to the specified widget.

Event filters

Mouse filters

enter :: Reactive w => Event w (Point -> IO ())

leave :: Reactive w => Event w (Point -> IO ())

motion :: Reactive w => Event w (Point -> IO ())

drag :: Reactive w => Event w (Point -> IO ())

click :: Reactive w => Event w (Point -> IO ())

unclick :: Reactive w => Event w (Point -> IO ())

doubleClick :: Reactive w => Event w (Point -> IO ())

clickRight :: Reactive w => Event w (Point -> IO ())

unclickRight :: Reactive w => Event w (Point -> IO ())

Keyboard event filters

anyKey :: Reactive w => Event w (Key -> IO ())

key :: Reactive w => Key -> Event w (IO ())

charKey :: Reactive w => Char -> Event w (IO ())

enterKey :: Reactive w => Event w (IO ())

tabKey :: Reactive w => Event w (IO ())

escKey :: Reactive w => Event w (IO ())

helpKey :: Reactive w => Event w (IO ())

delKey :: Reactive w => Event w (IO ())

homeKey :: Reactive w => Event w (IO ())

endKey :: Reactive w => Event w (IO ())

pgupKey :: Reactive w => Event w (IO ())

pgdownKey :: Reactive w => Event w (IO ())

downKey :: Reactive w => Event w (IO ())

upKey :: Reactive w => Event w (IO ())

leftKey :: Reactive w => Event w (IO ())

rightKey :: Reactive w => Event w (IO ())

rebind :: Event w (IO ()) -> Event w (IO ())

Types

Modifiers

data Modifiers :: *

The Modifiers indicate the meta keys that have been pressed (True) or not (False).

Constructors

Modifiers

Fields

altDown :: ~Bool

alt key down

shiftDown :: ~Bool

shift key down

controlDown :: ~Bool

control key down

metaDown :: ~Bool

meta key down

showModifiers :: Modifiers -> String

Show modifiers, for example for use in menus.

noneDown :: Modifiers

Construct a Modifiers structure with no meta keys pressed.

justShift :: Modifiers

Construct a Modifiers structure with just Shift meta key pressed.

justAlt :: Modifiers

Construct a Modifiers structure with just Alt meta key pressed.

justControl :: Modifiers

Construct a Modifiers structure with just Ctrl meta key pressed.

justMeta :: Modifiers

Construct a Modifiers structure with just Meta meta key pressed.

isNoneDown :: Modifiers -> Bool

Test if no meta key was pressed.

isNoShiftAltControlDown :: Modifiers -> Bool

Test if no shift, alt, or control key was pressed.

Mouse events

data EventMouse :: *

Mouse events. The Point gives the logical (unscrolled) position.

Constructors

MouseMotion ~Point ~Modifiers
Mouse was moved over the client area of the window

MouseEnter ~Point ~Modifiers
Mouse enters in the client area of the window

MouseLeave ~Point ~Modifiers
Mouse leaves the client area of the window

MouseLeftDown ~Point ~Modifiers
Mouse left button goes down

MouseLeftUp ~Point ~Modifiers
Mouse left button goes up

MouseLeftDClick ~Point ~Modifiers
Mouse left button double click

MouseLeftDrag ~Point ~Modifiers
Mouse left button drag

MouseRightDown ~Point ~Modifiers
Mouse right button goes down

MouseRightUp ~Point ~Modifiers
Mouse right button goes up

MouseRightDClick ~Point ~Modifiers
Mouse right button double click

MouseRightDrag ~Point ~Modifiers
Mouse right button drag (unsupported on most platforms)

MouseMiddleDown ~Point ~Modifiers
Mouse middle button goes down

MouseMiddleUp ~Point ~Modifiers
Mouse middle button goes up

MouseMiddleDClick ~Point ~Modifiers
Mouse middle button double click

MouseMiddleDrag ~Point ~Modifiers
Mouse middle button drag (unsupported on most platforms)

MouseWheel ~Bool ~Point ~Modifiers
Mouse wheel rotation. (Bool is True for a downward rotation)

showMouse :: EventMouse -> String

Show an EventMouse in a user friendly way.

mousePos :: EventMouse -> Point

Extract the position from a MouseEvent.

mouseModifiers :: EventMouse -> Modifiers

Extract the modifiers from a MouseEvent.

Calender event

data EventCalendar :: *

Constructors

CalendarDayChanged (DateTime ())

CalendarDoubleClicked (DateTime ())

CalendarMonthChanged (DateTime ())

CalendarSelectionChanged (DateTime ())

CalendarWeekdayClicked Int

CalendarYearChanged (DateTime ())

CalendarUnknown

calendarEvent :: Event (CalendarCtrl a) (EventCalendar -> IO ())

AuiNotebook event

data EventAuiNotebook :: *

AuiNotebook events.

Constructors

AuiNotebookAllowDnd

Fields

newSel :: WindowSelection

oldSel :: WindowSelection

AuiNotebookBeginDrag

Fields

newSel :: WindowSelection

oldSel :: WindowSelection

AuiNotebookBgDclick

Fields

newSel :: WindowSelection

oldSel :: WindowSelection

AuiNotebookButton

Fields

newSel :: WindowSelection

oldSel :: WindowSelection

AuiNotebookDragDone

Fields

newSel :: WindowSelection

oldSel :: WindowSelection

AuiNotebookDragMotion

Fields

newSel :: WindowSelection

oldSel :: WindowSelection

AuiNotebookEndDrag

Fields

newSel :: WindowSelection

oldSel :: WindowSelection

AuiNotebookPageChanged

Fields

newSel :: WindowSelection

oldSel :: WindowSelection

AuiNotebookPageChanging

Fields

newSel :: WindowSelection

oldSel :: WindowSelection

AuiNotebookPageClose

Fields

newSel :: WindowSelection

oldSel :: WindowSelection

AuiNotebookPageClosed

Fields

newSel :: WindowSelection

oldSel :: WindowSelection

AuiNotebookTabMiddleDown

Fields

newSel :: WindowSelection

oldSel :: WindowSelection

AuiNotebookTabMiddleUp

Fields

newSel :: WindowSelection

oldSel :: WindowSelection

AuiNotebookTabRightDown

Fields

newSel :: WindowSelection

oldSel :: WindowSelection

AuiNotebookTabRightUp

Fields

newSel :: WindowSelection

oldSel :: WindowSelection

AuiNotebookUnknown

AuiTabCtrlPageChanging

Fields

newSel :: WindowSelection

oldSel :: WindowSelection

AuiTabCtrlUnknown

auiNotebookOnPageCloseEvent :: Event (AuiNotebook a) (EventAuiNotebook -> IO ())

auiNotebookOnPageClosedEvent :: Event (AuiNotebook a) (EventAuiNotebook -> IO ())

auiNotebookOnPageChangingEvent :: Event (AuiNotebook a) (EventAuiNotebook -> IO ())

auiNotebookOnPageChangedEvent :: Event (AuiNotebook a) (EventAuiNotebook -> IO ())

Keyboard events

data EventKey :: *

A keyboard event contains the key, the modifiers and the focus point.

Constructors

EventKey ~Key ~Modifiers ~Point

data Key :: *

A Key represents a single key on a keyboard.

Constructors

KeyChar ~Char
An ascii code.

KeyOther ~KeyCode
An unknown virtual key.

KeyBack

KeyTab

KeyReturn

KeyEscape

KeySpace

KeyDelete

KeyInsert

KeyEnd

KeyHome

KeyLeft

KeyUp

KeyRight

KeyDown

KeyPageUp

KeyPageDown

KeyStart

KeyClear

KeyShift

KeyAlt

KeyControl

KeyMenu

KeyPause

KeyCapital

KeyHelp

KeySelect

KeyPrint

KeyExecute

KeySnapshot

KeyCancel

KeyLeftButton

KeyRightButton

KeyMiddleButton

KeyNum0

KeyNum1

KeyNum2

KeyNum3

KeyNum4

KeyNum5

KeyNum6

KeyNum7

KeyNum8

KeyNum9

KeyMultiply

KeyAdd

KeySeparator

KeySubtract

KeyDecimal

KeyDivide

KeyF1

KeyF2

KeyF3

KeyF4

KeyF5

KeyF6

KeyF7

KeyF8

KeyF9

KeyF10

KeyF11

KeyF12

KeyF13

KeyF14

KeyF15

KeyF16

KeyF17

KeyF18

KeyF19

KeyF20

KeyF21

KeyF22

KeyF23

KeyF24

KeyNumLock

KeyScroll

keyKey :: EventKey -> Key

Extract the key from a keyboard event.

keyModifiers :: EventKey -> Modifiers

Extract the modifiers from a keyboard event.

keyPos :: EventKey -> Point

Extract the position from a keyboard event.

showKey :: Key -> String

Show a key for use in menus for example.

showKeyModifiers :: Key -> Modifiers -> String

Show a key/modifiers combination, for example for use in menus.

Internal

newEvent :: String -> (w -> IO a) -> (w -> a -> IO ()) -> Event w a

Create a new event from a get and set function.