Graphics.UI.WX.Frame

Description

Frames.

Instances: Form, Closeable, Framed -- Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint, HasDefault.

Frames

type Frame a = TopLevelWindow (CFrame a)

Pointer to an object of type Frame, derived from TopLevelWindow.

frame :: [Prop (Frame ())] -> IO (Frame ())

Create a top-level frame window.

frameFixed :: [Prop (Frame ())] -> IO (Frame ())

Create a top-level frame window that is not resizeable.

frameTool :: [Prop (Frame ())] -> Window a -> IO (Frame ())

Create a tool window; floats on the parent and has a small caption.

frameEx :: Style -> [Prop (Frame ())] -> Window a -> IO (Frame ())

Create a top-level frame window in a custom style.

frameLoadRes :: FilePath -> String -> [Prop (Frame ())] -> IO (Frame ())

Complete the construction of a top level frame which has been loaded from a resource file.

frameLoadChildRes :: Window a -> FilePath -> String -> [Prop (Frame ())] -> IO (Frame ())

Complete the construction of a frame which is the child of some existing parent window.

initialFrame :: (Id -> Rect -> String -> [Prop (Window w)] -> Style -> a) -> [Prop (Window w)] -> Style -> a

initial Frame flags

MDI Frames

type MDIParentFrame a = Frame (CMDIParentFrame a)

Pointer to an object of type MDIParentFrame, derived from Frame.

type MDIChildFrame a = Frame (CMDIChildFrame a)

Pointer to an object of type MDIChildFrame, derived from Frame.

mdiParentFrame :: [Prop (MDIParentFrame ())] -> IO (MDIParentFrame ())

Create an MDI parent frame.

mdiChildFrame :: MDIParentFrame a -> [Prop (MDIChildFrame ())] -> IO (MDIChildFrame ())

Create a MDI child frame.

mdiParentFrameEx :: Window a -> Style -> [Prop (MDIParentFrame ())] -> IO (MDIParentFrame ())

Create an MDI parent frame with a custom style.

mdiChildFrameEx :: MDIParentFrame a -> Style -> [Prop (MDIChildFrame ())] -> IO (MDIChildFrame ())

Create a MDI child frame with a custom style.

Operations

activeChild :: ReadAttr (MDIParentFrame a) (MDIChildFrame ())

Return the active child frame (objectIsNull when no child is active)

activateNext :: MDIParentFrame a -> IO ()

Activate the next child frame.

activatePrevious :: MDIParentFrame a -> IO ()

Activate the previous child frame

arrangeIcons :: MDIParentFrame a -> IO ()

Arrange iconized mdi child frames.

cascade :: MDIParentFrame a -> IO ()

Cascade the child frames.

tile :: MDIParentFrame a -> IO ()

Tile the child frames

Internal