Graphics.UI.WX.Layout

Description

Just re-exports functionality of Graphics.UI.WXCore.Layout. See that module for a description of layout combinators.

Any object in the Form class has a layout attribute to specify the layout. Here is a short example:

do f <- frame [text := "layout demo"]
   q <- button f [text := "quit", on command := close f]
   set f [layout := margin 10 $
                    floatCentre $
                    column 5 [label "hi",widget q]]

Classes

class Form w where

Forms can contain other widgets.

Minimal complete definition

layout

Methods

layout :: Attr w Layout

Set the layout of the child widgets.

layout :: Form w => Attr w Layout

Set the layout of the child widgets.

Layout

module Graphics.UI.WXCore.Layout