Graphics.UI.WX.Media

Description

Images, Media, Sounds, and action!

Media

class Media w where

Abstract layer between MediaCtrl and Sound. This class intends to avoid breaking backward-compatibility.

Minimal complete definition

play, stop

Methods

play :: w -> IO ()

If use this method with Sound, play a sound fragment asynchronously. If use this method with MediaCtrl, play media that is loaded by mediaCtrlLoad.

stop :: w -> IO ()

Sound

sound :: FilePath -> Sound ()

Return a managed sound object. The file path points to a valid sound file, normally a .wav.

playLoop :: Sound a -> IO ()

Play a sound fragment repeatedly (and asynchronously).

playWait :: Sound a -> IO ()

Play a sound fragment synchronously (i.e. wait till completion).

Images

image :: FilePath -> Image ()

Return a managed image. Images are platform independent representations of pictures, using an array of rgb pixels. See Graphics.UI.WXCore.Image for low-level pixel manipulation. The file path should point to a valid image file, like .jpg, .bmp, .xpm, or .png, for example.

Instances: Sized.

imageCreateFromFile :: String -> IO (Image ())

usage: (imageCreateFromFile name).

imageCreateFromPixels :: Size -> [Color] -> IO (Image ())

Create an image from a list of pixels.

imageGetPixels :: Image a -> IO [Color]

Get the pixels of an image.

imageCreateFromPixelArray :: IArray a Color => a Point Color -> IO (Image ())

Create an image from a pixel array

imageGetPixelArray :: IArray a Color => Image b -> IO (a Point Color)

Get the pixels of an image as an array

Bitmaps

bitmap :: FilePath -> Bitmap ()

Return a managed bitmap object. Bitmaps are abstract images used for drawing to a device context. The file path should point to a valid bitmap file, normally a .ico, .bmp, .xpm, or .png, but any file format supported by Image is correctly loaded.

Instances: Sized.

bitmapCreateFromFile :: FilePath -> IO (Bitmap ())

Load a bitmap from an image file (gif, jpg, png, etc.)

bitmapFromImage :: Image a -> IO (Bitmap ())

Create a bitmap from an image with the same color depth.

Orphan instances

Sized (Bitmap a)

Methods

size :: Attr (Bitmap a) Size

Sized (Image a)

Methods

size :: Attr (Image a) Size