Graphics.UI.WX.Timer

Description

Support for millisecond timers.

Documentation

type Timer = TimerEx ()

A timer generates a command event on a specified millisecond interval.

Attributes: interval

Instances: Able, Commanding

timer :: Window a -> [Prop Timer] -> IO Timer

Create a new timer with a 1 second interval. The timer is automatically discarded when the parent is deleted.

interval :: Attr Timer Int

The millisecond interval of the timer.

Orphan instances

Able Timer

Methods

enabled :: Attr Timer Bool

Commanding Timer

Methods

command :: Event Timer (IO ())