Graphics.UI.WX.TopLevelWindow

Description

wxTopLevelwindow (wxWidgets >= 2.8.0) defines an (abstract) common base class for wxFrame and wxDialog.

In the wxHaskell implementation, TopLevel has been added to encapsulate some of the common functionality between the Dialog and Frame modules.

Instances: HasDefault, Pictured, Framed, Form, Closeable

Instances inherited from Window: Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

Documentation

initialResizeable :: ([Prop (Window w)] -> Style -> a) -> [Prop (Window w)] -> Style -> a

Helper function that transforms the style accordding to the windowResizable flag in of the properties

initialMaximizeable :: ([Prop (Window w)] -> Style -> a) -> [Prop (Window w)] -> Style -> a

Helper function that transforms the style accordding to the windowMaximizable flag in of the properties

initialMinimizeable :: ([Prop (Window w)] -> Style -> a) -> [Prop (Window w)] -> Style -> a

Helper function that transforms the style accordding to the windowMinimizable flag in of the properties

initialCloseable :: ([Prop (Window w)] -> Style -> a) -> [Prop (Window w)] -> Style -> a

Helper function that transforms the style accordding to the windowMinimizable flag in of the properties

Orphan instances

Form (Frame a)

Methods

layout :: Attr (Frame a) Layout

HasDefault (TopLevelWindow a)

Methods

unsafeDefaultItem :: Attr (TopLevelWindow a) (Window ())

defaultButton :: Attr (TopLevelWindow a) (Button ())

Framed (TopLevelWindow a)

Methods

resizeable :: Attr (TopLevelWindow a) Bool

minimizeable :: Attr (TopLevelWindow a) Bool

maximizeable :: Attr (TopLevelWindow a) Bool

closeable :: Attr (TopLevelWindow a) Bool

Closeable (Frame a)

Methods

close :: Frame a -> IO ()

Pictured (TopLevelWindow a)

Methods

picture :: Attr (TopLevelWindow a) FilePath