Graphics.UI.WX.Window

Description

Exports default instances for generic windows.

Instances: Textual, Literate, Dimensions, Colored, Visible, Child, Sized, Parent, Help, Bordered, Able, Tipped, Identity, Styled, Reactive, Paint.

Window

type Window a = EvtHandler (CWindow a)

Pointer to an object of type Window, derived from EvtHandler.

window :: Window a -> [Prop (Window ())] -> IO (Window ())

Create a plain window. Can be used to define custom controls for example.

Attributes: rootParent, frameParent, tabTraversal

Instances: Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

refit :: Window a -> IO ()

Ensure that a widget is refitted inside a window when its size changes, for example when the text of a staticText control changes. (Calls windowReFit)

refitMinimal :: Window a -> IO ()

Ensure that a widget is refitted inside a window when its size changes, for example when the text of a staticText control changes. Always resizes the window to its minimal acceptable size. (Calls windowReFitMinimal)

rootParent :: ReadAttr (Window a) (Window ())

The ultimate root parent of the widget.

frameParent :: ReadAttr (Window a) (Window ())

The parent frame or dialog of a widget.

tabTraversal :: Attr (Window a) Bool

Enable (or disable) tab-traversal. (= wxTAB_TRAVERSAL)

ScrolledWindow

type ScrolledWindow a = Panel (CScrolledWindow a)

Pointer to an object of type ScrolledWindow, derived from Panel.

scrolledWindow :: Window a -> [Prop (ScrolledWindow ())] -> IO (ScrolledWindow ())

A scrollable window. Use virtualSize and scrollRate to set the scrollbar behaviour.

Attributes: scrollRate

Instances: Textual, Literate, Dimensions, Colored, Visible, Child, Able, Tipped, Identity, Styled, Reactive, Paint.

scrollRate :: Attr (ScrolledWindow a) Size

The horizontal and vertical scroll rate of scrolled window. Use 0 to disable scrolling in that direction.

Internal

initialWindow :: (Id -> Rect -> [Prop (Window w)] -> Style -> a) -> [Prop (Window w)] -> Style -> a

Helper function that retrieves initial window settings, including |identity|, |style|, and |area| (or |position| and |outerSize|).

initialContainer :: (Id -> Rect -> [Prop (Window w)] -> Style -> a) -> [Prop (Window w)] -> Style -> a

Helper function that retrieves initial window settings, including |clipChildren| and |fullRepaintOnResize|.

initialIdentity :: Identity w => (Id -> [Prop w] -> a) -> [Prop w] -> a

Helper function that retrieves the initial |identity|.

initialStyle :: Styled w => ([Prop w] -> Style -> a) -> [Prop w] -> Style -> a

Helper function that retrieves the initial |style|.

initialText :: Textual w => (String -> [Prop w] -> a) -> [Prop w] -> a

Retrieve the initial title from the |text| attribute.

initialFullRepaintOnResize :: Visible w => ([Prop w] -> Style -> a) -> [Prop w] -> Style -> a

Helper function that transforms the style accordding to the fullRepaintOnResize flag in of the properties

initialClipChildren :: Parent w => ([Prop w] -> Style -> a) -> [Prop w] -> Style -> a

Helper function that transforms the style accordding to the clipChildren flag out of the properties

Orphan instances

Sized (Window a)

Methods

size :: Attr (Window a) Size

Tipped (Window a)

Methods

tooltip :: Attr (Window a) String

Styled (Window a)

Methods

style :: Attr (Window a) Int

Identity (Window a)

Methods

identity :: Attr (Window a) Int

Able (Window a)

Methods

enabled :: Attr (Window a) Bool

Child (Window a)

Methods

parent :: ReadAttr (Window a) (Window ())

Bordered (Window a)

Methods

border :: Attr (Window a) Border

Parent (Window a)

Methods

children :: ReadAttr (Window a) [Window ()]

clipChildren :: Attr (Window a) Bool

Visible (Window a)

Methods

visible :: Attr (Window a) Bool

refresh :: Window a -> IO ()

fullRepaintOnResize :: Attr (Window a) Bool

Colored (Window a)

Methods

bgcolor :: Attr (Window a) Color

color :: Attr (Window a) Color

Dimensions (Window a)

Methods

outerSize :: Attr (Window a) Size

position :: Attr (Window a) Point

area :: Attr (Window a) Rect

bestSize :: ReadAttr (Window a) Size

clientSize :: Attr (Window a) Size

virtualSize :: Attr (Window a) Size

Literate (Window a)

Methods

font :: Attr (Window a) FontStyle

fontSize :: Attr (Window a) Int

fontWeight :: Attr (Window a) FontWeight

fontFamily :: Attr (Window a) FontFamily

fontShape :: Attr (Window a) FontShape

fontFace :: Attr (Window a) String

fontUnderline :: Attr (Window a) Bool

textColor :: Attr (Window a) Color

textBgcolor :: Attr (Window a) Color

Textual (Window a)

Methods

text :: Attr (Window a) String

appendText :: Window a -> String -> IO ()

Paint (Window a)

Methods

paint :: Event (Window a) (DC () -> Rect -> IO ())

paintRaw :: Event (Window a) (PaintDC () -> Rect -> [Rect] -> IO ())

paintGc :: Event (Window a) (GCDC () -> Rect -> IO ())

repaint :: Window a -> IO ()

Reactive (Window a)

Methods

mouse :: Event (Window a) (EventMouse -> IO ())

keyboard :: Event (Window a) (EventKey -> IO ())

closing :: Event (Window a) (IO ())

idle :: Event (Window a) (IO Bool)

resize :: Event (Window a) (IO ())

focus :: Event (Window a) (Bool -> IO ())

activate :: Event (Window a) (Bool -> IO ())